feat: rename FileDropInput to MediaFileDropInput and enhance media file handling with type-specific icons
