+++ apps/ui/app/components/media-file-drop-input.tsx
@@ -1,4 +1,12 @@
-import { CloudUploadIcon, XIcon } from 'lucide-react'
+import {
+  CloudUploadIcon,
+  FileAudio,
+  FileIcon,
+  FileImage,
+  FileVideo,
+  PlayIcon,
+  XIcon,
+} from 'lucide-react'
 import { cn } from '~/lib/utils'
 import { FileDrop } from './file-drop'
 import { Button } from './ui'
@@ -7,6 +15,19 @@ const accepts = {
   image: ['.png', '.jpg', '.jpeg', '.gif', '.webp'],
   video: ['.mp4', '.webm'],
   audio: ['.mp3', '.ogg', '.wav'],
+}
+
+const MediaIcon = ({ type }: { type: string }) => {
+  if (type.startsWith('image')) {
+    return <FileImage className="h-6 w-6" />
+  }
+  if (type.startsWith('video')) {
+    return <FileVideo className="h-6 w-6" />
+  }
+  if (type.startsWith('audio')) {
+    return <FileAudio className="h-6 w-6" />
+  }
+  return <FileIcon className="h-6 w-6" />
 }
 
 export const MediaFileDropInput = ({
@@ -34,13 +55,31 @@ export const MediaFileDropInput = ({
       {({ isDragging, files, removeFile }) => (
         <div className="flex flex-col items-center gap-4 text-center">
           <CloudUploadIcon className="size-12 stroke-muted-foreground" />
+
           {isDragging ? (
             <p>ファイルをここにドロップ</p>
           ) : (
             <>
-              <p className="italic text-muted-foreground hover:text-accent-foreground">
-                <strong>{files.map((f) => f.name)}</strong>
-              </p>
+              {files.map((f, index) => (
+                <div
+                  key={`${f.name}_${index}`}
+                  className="grid w-full grid-cols-[auto_1fr_auto] place-items-center gap-4 italic text-muted-foreground"
+                >
+                  <MediaIcon type={f.type} />
+                  <strong>{files.map((f) => f.name)}</strong>
+                  <Button
+                    type="button"
+                    variant="outline"
+                    size="sm"
+                    onClick={(e) => {
+                      e.stopPropagation()
+                    }}
+                  >
+                    <PlayIcon />
+                  </Button>
+                </div>
+              ))}
+
               {files.length >= 1 && (
                 <Button
                   type="button"
