--- apps/ui/app/routes/_app+/advertiser+/entries+/new/components/index.ts
@@ -1,3 +1,0 @@
-export * from './ad-form'
-export * from './ad-group-form'
-export * from './campaign-form'
